Move fiscalYear prompt to reflect it is no longer a followUp step
