+++ src/workers/extractEmissions.ts
@@ -7,7 +7,6 @@ import scope3 from '../prompts/followUp/scope3'
 import goals from '../prompts/followUp/goals'
 import initiatives from '../prompts/followUp/initiatives'
 import baseFacts from '../prompts/followUp/baseFacts'
-import fiscalYear from '../prompts/followUp/fiscalYear'
 import { DiscordJob, DiscordWorker } from '../lib/DiscordWorker'
 import biogenic from '../prompts/followUp/biogenic'
